Move everything else to 'common' package
