+++ eslint.config.js
@@ -26,8 +26,6 @@ export default tseslint.config([
 
             // 3. Internal aliases (custom order)
             ["^@components"],
-            ["^@models"],
-            ["^@config"],
 
             // 4. CSS/SCSS imports
             ["^.+\\.css$", "^.+\\.scss$"],
