+++ src/components/products/category.service.ts
@@ -1,8 +1,7 @@
 import axios from "axios";
 
+import { appConfig } from "@components/common/app.config";
 import type { Category } from "@components/products/category";
-
-import { appConfig } from "@config/app.config";
 
 const getCategories = async (): Promise<Category[]> => {
   const { data } = await axios.get<Category[]>(
