+++ src/components/products/product.service.ts
@@ -1,8 +1,7 @@
 import axios from "axios";
 
+import { appConfig } from "@components/common/app.config";
 import type { Product } from "@components/products/product";
-
-import { appConfig } from "@config/app.config";
 
 const getProducts = async (categoryId?: number): Promise<Product[]> => {
   let endpoint = `${appConfig.backendUrl}/products`;
